refactor(harness): split agent_loop/mod.rs into entry/run_loop/model_call
agent_loop/mod.rs had grown to ~1160 lines mixing public invoke*
entry points, the ~400-line run_loop body, and model-invocation
dispatch in one impl block. Split mechanically into entry.rs
(invoke*/invoke_streaming*/drive), run_loop.rs (the core loop body
and response-cache decision), and model_call.rs (cache-aware retry/
fallback dispatch plus the ModelBaseCall/ToolBaseCall impls). No
behavior change.
